EditorButtonContent component added.
